LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-inactive-expected.txt


FAIL Animation does not apply when the timeline is inactive because there is not scroll range assert_equals: expected "0px" but got "100px"
FAIL Animation does not apply when timeline is initially inactive assert_equals: expected "0px" but got "100px"
